Moved crypto typings into src
