+++ lib/packages/crypto/src/decryptData.ts
@@ -9,7 +9,7 @@ import * as CryptoJS from 'crypto-js';
 import {inflate} from 'pako/lib/inflate';
 import {ECKCDSA} from './ec-kcdsa';
 import {Converter} from './converter';
-import {EncryptedData} from '../typings/encryptedData';
+import {EncryptedData} from './typings/encryptedData';
 
 const IV_LENGTH = 16;
 const SHARED_KEY_SIZE = 32;
