+++ lib/packages/crypto/src/encryptData.ts
@@ -9,7 +9,7 @@ import * as CryptoJS from 'crypto-js';
 import {gzip} from 'pako/lib/deflate';
 import {ECKCDSA} from './ec-kcdsa';
 import {Converter} from './converter';
-import {EncryptedData} from '../typings/encryptedData';
+import {EncryptedData} from './typings/encryptedData';
 
 const IV_LENGTH = 16;
 const SHARED_KEY_SIZE = 32;
