+++ lib/packages/crypto/src/index.ts
@@ -14,6 +14,6 @@ export * from './passPhraseGenerator';
 export * from './verifySignature';
 export * from './generateSignedTransactionBytes';
 export * from './generateSignature';
-export * from '../typings/keys';
-export * from '../typings/encryptedMessage';
-export * from '../typings/encryptedData';
+export * from './typings/keys';
+export * from './typings/encryptedMessage';
+export * from './typings/encryptedData';
