+++ lib/packages/crypto/src/typings/encryptedData.ts
