+++ lib/packages/crypto/src/typings/encryptedMessage.ts
